Time: 5 ms (38.14%), Space: 69.1 MB (95.46%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.java
@@ -14,6 +14,6 @@ class Solution {
         j++;
        }
       
-       return (double)(maxSum / k);
+       return (maxSum / k);
     }
 }
